backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djoser==2.3.0
Pillow==11.0.0
django-filter==24.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
